Fix the build of docker-compose
The cloudera archive requires now authentication.
Effective Jan 31, 2021, all Cloudera software requires a subscription.
Put an older version of CDH tarball hadoop-2.6.0-cdh5.11.0 directly in the repo,
but user should ask to cloudera. It could be improved in a second commit, by
building a garmadon docker image directly and put it in docker hub.

In addition, cleaned set-up for hadoop 3, before re-doing it, because it
doesn't work more.

Add .gitattributes for other OS support: Windows, macOS to have
end-of-line as Linux, and know binary files.